refactor(dashboard): reorganización de las páginas de perfil y del panel de control
- Las páginas del panel de control relacionadas con el usuario han sido movidas a /(dashboard)/user/.
- Se ha creado una nueva página de perfil para refugios en /(dashboard)/shelter/profile.
- La página de perfil de administrador ha sido actualizada a /(dashboard)/admin/profile.
- La lógica de redireccionamiento y el enrutamiento para los diferentes roles de usuario han sido ajustados.
- Los metadatos y los enlaces de navegación (breadcrumbs) han sido actualizados para una mejor navegación.
