+++ CAMBIOS_ESTRUCTURA.md
@@ -1,0 +1,216 @@
+# ✅ CAMBIOS ESTRUCTURALES APLICADOS - PawLig
+
+**Fecha:** 2025  
+**Responsable:** Equipo de Desarrollo  
+**Estado:** COMPLETADO
+
+---
+
+## 📋 RESUMEN DE CAMBIOS
+
+### ❌ ELIMINADOS (Rutas incorrectas)
+
+1. **`app/(dashboard)/profile/page.tsx`** ❌ ELIMINADO
+   - **Razón:** Ruta genérica incorrecta. Cada rol tiene su propia ruta de perfil.
+   - **Impacto:** Ninguno (no se usaba en producción)
+
+2. **`app/(dashboard)/adopter/`** ❌ ELIMINADO (carpeta completa)
+   - **Razón:** Nombre incorrecto según arquitectura. El rol ADOPTER usa `/user`, no `/adopter`.
+   - **Contenía:** `adopter/profile/page.tsx`
+   - **Migrado a:** `user/page.tsx` y `user/profile/page.tsx`
+
+---
+
+## ✅ CREADOS (Rutas correctas)
+
+### 1. **`app/(dashboard)/user/page.tsx`** ✅ CREADO
+   - **Propósito:** Dashboard principal del usuario adoptante
+   - **Ruta final:** `/user`
+   - **Requerimiento:** HU-004 - Visualización del Panel de Usuario
+   - **Funcionalidad:**
+     - Ver mascotas favoritas
+     - Ver estado de solicitudes de adopción
+     - Notificaciones de cambios
+   - **Componente:** `AdopterDashboardClient`
+
+### 2. **`app/(dashboard)/user/profile/page.tsx`** ✅ CREADO
+   - **Propósito:** Edición de perfil personal del adoptante
+   - **Ruta final:** `/user/profile`
+   - **Requerimiento:** RF-003 - Actualización del perfil del usuario
+   - **Funcionalidad:**
+     - Editar información personal
+     - Validación de campos obligatorios
+     - Guardado inmediato
+   - **Componente:** `UserProfileForm`
+
+### 3. **`app/(dashboard)/shelter/profile/page.tsx`** ✅ CREADO
+   - **Propósito:** Edición de perfil del albergue
+   - **Ruta final:** `/shelter/profile`
+   - **Requerimiento:** HU-003 para SHELTER
+   - **Estado:** Estructura base creada (funcionalidad en desarrollo)
+
+### 4. **`app/(dashboard)/admin/profile/page.tsx`** ✅ CREADO
+   - **Propósito:** Edición de perfil del administrador
+   - **Ruta final:** `/admin/profile`
+   - **Estado:** Estructura base creada (funcionalidad en desarrollo)
+
+---
+
+## ✅ MANTENIDOS (Rutas correctas existentes)
+
+- ✅ `app/(dashboard)/admin/users/page.tsx` - Gestión de usuarios
+- ✅ `app/(dashboard)/vendor/profile/page.tsx` - Perfil de vendedor
+- ✅ `app/(dashboard)/shelter/pets/page.tsx` - Gestión de mascotas
+- ✅ `app/(dashboard)/shelter/pets/new/page.tsx` - Crear mascota
+- ✅ `app/(dashboard)/shelter/pets/[id]/edit/page.tsx` - Editar mascota
+- ✅ `app/(dashboard)/shelter/adoptions/page.tsx` - Postulaciones
+
+---
+
+## 🗂️ ESTRUCTURA FINAL DEL PROYECTO
+
+```
+app/(dashboard)/
+├── admin/
+│   ├── profile/
+│   │   └── page.tsx          ✅ NUEVO - Perfil de admin
+│   └── users/
+│       ├── page.tsx          ✅ Gestión de usuarios
+│       ├── UsersManagementClient.tsx
+│       └── BlockUserModal.tsx
+│
+├── shelter/
+│   ├── profile/
+│   │   └── page.tsx          ✅ NUEVO - Perfil de albergue
+│   ├── adoptions/
+│   │   └── page.tsx          ✅ Postulaciones
+│   └── pets/
+│       ├── page.tsx          ✅ Lista de mascotas
+│       ├── new/
+│       │   └── page.tsx      ✅ Crear mascota
+│       └── [id]/
+│           └── edit/
+│               └── page.tsx  ✅ Editar mascota
+│
+├── user/                      ✅ NUEVO - Reemplaza /adopter
+│   ├── page.tsx              ✅ NUEVO - Dashboard de adoptante
+│   └── profile/
+│       └── page.tsx          ✅ NUEVO - Perfil de adoptante
+│
+└── vendor/
+    └── profile/
+        └── page.tsx          ✅ Perfil de vendedor
+```
+
+---
+
+## 🔐 PROTECCIÓN DE RUTAS (middleware.ts)
+
+El middleware ya está configurado correctamente:
+
+```typescript
+export const config = {
+  matcher: [
+    "/request-shelter",
+    "/admin/:path*",      // Solo ADMIN
+    "/shelter/:path*",    // Solo SHELTER
+    "/vendor/:path*",     // Solo VENDOR
+    "/user/:path*",       // ✅ Ya configurado para ADOPTER
+  ],
+};
+```
+
+---
+
+## 📊 CHECKLIST DE VALIDACIÓN
+
+- ✅ Carpeta `adopter/` renombrada a `user/`
+- ✅ Archivo `/profile/page.tsx` eliminado
+- ✅ Dashboard de usuario creado en `/user/page.tsx`
+- ✅ Perfil de usuario creado en `/user/profile/page.tsx`
+- ✅ Perfil de albergue creado en `/shelter/profile/page.tsx`
+- ✅ Perfil de admin creado en `/admin/profile/page.tsx`
+- ✅ Todas las rutas protegidas por middleware
+- ✅ Componentes existentes compatibles (AdopterDashboardClient, UserProfileForm)
+- ✅ Sin referencias a rutas antiguas en código fuente
+
+---
+
+## 🧪 PRUEBAS RECOMENDADAS
+
+### 1. Navegación de Usuario ADOPTER
+```
+1. Login como ADOPTER
+2. Acceder a /user → Debe mostrar dashboard
+3. Acceder a /user/profile → Debe mostrar formulario de perfil
+4. Intentar acceder a /adopter → Debe dar 404
+5. Intentar acceder a /profile → Debe dar 404
+```
+
+### 2. Navegación de Usuario SHELTER
+```
+1. Login como SHELTER
+2. Acceder a /shelter/profile → Debe mostrar perfil
+3. Acceder a /shelter/pets → Debe mostrar mascotas
+4. Intentar acceder a /user → Debe redirigir a /unauthorized
+```
+
+### 3. Navegación de Usuario ADMIN
+```
+1. Login como ADMIN
+2. Acceder a /admin/profile → Debe mostrar perfil
+3. Acceder a /admin/users → Debe mostrar gestión
+4. Acceder a /user → Debe permitir (ADMIN tiene acceso total)
+```
+
+### 4. Navegación de Usuario VENDOR
+```
+1. Login como VENDOR
+2. Acceder a /vendor/profile → Debe mostrar perfil
+3. Intentar acceder a /user → Debe redirigir a /unauthorized
+```
+
+---
+
+## 🔄 PRÓXIMOS PASOS
+
+1. **Implementar formularios completos:**
+   - `/shelter/profile/page.tsx` → Crear `ShelterProfileForm`
+   - `/admin/profile/page.tsx` → Crear `AdminProfileForm`
+
+2. **Actualizar enlaces en navegación:**
+   - Verificar que todos los links apunten a `/user` en lugar de `/adopter`
+   - Actualizar breadcrumbs si existen
+
+3. **Limpiar build:**
+   ```bash
+   npm run build
+   ```
+
+4. **Ejecutar pruebas:**
+   - Probar cada ruta con cada rol
+   - Verificar redirecciones del middleware
+   - Confirmar que formularios funcionan correctamente
+
+---
+
+## 📝 NOTAS TÉCNICAS
+
+- **Compatibilidad:** Todos los componentes existentes son compatibles con la nueva estructura
+- **Middleware:** Ya estaba configurado para `/user/:path*`, no requiere cambios
+- **Base de datos:** No requiere migraciones (solo cambios de rutas frontend)
+- **SEO:** Metadata actualizada en cada página nueva
+- **TypeScript:** Todos los archivos nuevos usan TypeScript estricto
+
+---
+
+## 👥 EQUIPO
+
+- **Andrés Ospina** - Líder y Desarrollador Backend
+- **Mateo Úsuga** - Desarrollador y Analista
+- **Santiago Lezcano** - Diseñador y Tester
+
+---
+
+**Documento generado automáticamente**  
+**PawLig - SENA 2025**
