+++ app/(dashboard)/shelter/profile/page.tsx
@@ -1,0 +1,60 @@
+import { Metadata } from 'next';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/lib/auth/auth-options';
+import { redirect } from 'next/navigation';
+
+export const metadata: Metadata = {
+  title: 'Editar Perfil del Albergue - PawLig',
+  description: 'Actualiza la información de tu albergue',
+};
+
+export default async function ShelterProfilePage() {
+  const session = await getServerSession(authOptions);
+
+  if (!session?.user) {
+    redirect('/login?callbackUrl=/shelter/profile');
+  }
+
+  if (session.user.role !== 'SHELTER') {
+    redirect('/unauthorized');
+  }
+
+  return (
+    <div className="min-h-screen bg-gray-50">
+      <header className="bg-white shadow-sm">
+        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-4">
+          <div className="flex justify-between items-center">
+            <h1 className="text-2xl font-bold text-purple-600">PawLig</h1>
+            <div className="flex items-center gap-4">
+              <span className="text-sm text-gray-600">
+                Hola, <span className="font-semibold">{session.user.name}</span>
+              </span>
+              <a href="/api/auth/signout" className="text-sm text-gray-600 hover:text-gray-900">
+                Cerrar sesión
+              </a>
+            </div>
+          </div>
+        </div>
+      </header>
+
+      <main className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
+        <div className="mb-8">
+          <a href="/shelter/pets" className="text-purple-600 hover:text-purple-700 text-sm font-semibold">
+            ← Volver a Mascotas
+          </a>
+        </div>
+
+        <div className="bg-white rounded-lg shadow-md p-8">
+          <h2 className="text-2xl font-bold text-gray-900 mb-6">Editar Perfil del Albergue</h2>
+          <p className="text-gray-600">Funcionalidad en desarrollo</p>
+        </div>
+      </main>
+
+      <footer className="bg-white border-t border-gray-200 mt-16">
+        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
+          <p className="text-center text-gray-500 text-sm">&copy; 2025 - PawLig<br />Todos los derechos reservados</p>
+        </div>
+      </footer>
+    </div>
+  );
+}
